refactor: rename itemstore to content
